convert usermanager to typescript
closes #901
